eval data string
size: 3828
bench: 6129131
