+++ src/eval.cpp
@@ -8,27 +8,8 @@ int VALUE[] = { 100, 320, 330, 500, 900, 2000, 0 };
 // Values copied from 4ku, will replace later
 int MATERIAL[] = { S(89, 147), S(350, 521), S(361, 521), S(479, 956), S(1046, 1782), 0 };
 
-int PST_RANK[] = {
-    0, S(-3, 0), S(-3, -1), S(-1, -1), S(1, 0), S(5, 3), 0, 0,
-    S(-2, -5), S(0, -3), S(1, -1), S(3, 3), S(4, 4), S(5, 1), S(2, 0), S(-15, 1),
-    S(0, -2), S(2, -1), S(2, 0), S(2, 0), S(2, 0), S(2, 0), S(-1, 0), S(-10, 2),
-    S(0, -3), S(-1, -3), S(-2, -2), S(-2, 0), S(0, 2), S(2, 2), S(1, 3), S(2, 1),
-    S(2, -11), S(3, -8), S(2, -3), S(0, 2), S(0, 5), S(-1, 5), S(-4, 7), S(-2, 4),
-    S(-1, -6), S(1, -2), S(-1, 0), S(-4, 3), S(-1, 5), S(5, 4), S(5, 2), S(5, -6),
-};
-
-int PST_FILE[] = {
-    S(-1, 1), S(-2, 1), S(-1, 0), S(0, -1), S(1, 0), S(2, 0), S(2, 0), S(-1, 0),
-    S(-4, -3), S(-1, -1), S(0, 1), S(2, 3), S(2, 3), S(2, 0), S(1, -1), S(-1, -3),
-    S(-2, -1), 0, S(1, 0), S(0, 1), S(1, 1), S(0, 1), S(2, 0), S(-1, -1),
-    S(-2, 0), S(-1, 1), S(0, 1), S(1, 0), S(2, -1), S(1, 0), S(1, 0), S(-1, -1),
-    S(-2, -3), S(-1, -1), S(-1, 0), S(0, 1), S(0, 2), S(1, 2), S(2, 0), S(1, -1),
-    S(-2, -5), S(2, -1), S(-1, 1), S(-4, 2), S(-4, 2), S(-2, 2), S(2, -1), S(0, -5),
-};
-
-int MOBILITY[] = { 0, S(8, 5), S(7, 7), S(3, 5), S(3, 2), S(-5, -1) };
-
-int PASSER[] = { 0, S(0, 10), S(-5, 20), S(-10, 30), S(10, 50), S(15, 100), S(25, 150), 0 };
+#define SCALE_PST 8
+#define SCALE_PASSER 5
 
 #define BISHOP_PAIR S(25, 55)
 #define KING_OPEN S(-75, 5)
@@ -37,3 +18,23 @@ int PASSER[] = { 0, S(0, 10), S(-5, 20), S(-10, 30), S(10, 50), S(15, 100), S(25
 #define ROOK_SEMI_OPEN S(10, 15)
 
 #define TEMPO 20
+
+#define DATA_STR "/,,.04//-/02341 /11111.%/.--/101121//.+-.0.+.444#\"#$%&&# #$&&&%#\"$%$%$&#\"#$%&%%#\"##$$%&%\"&#  \"&$%-,(( \"\"! $%'++**+.++&(*./,+,)*+++++-(()+--., #(-002/%)+.0/-%&&%$%%%%\"$&((%$\"$%%&&&%$%&&%$%%$\"$%&''%$ $&'''$ !&(&#  \"$&*4>"
+
+#define INDEX_EG 109
+
+#define INDEX_PST_RANK 0
+#define INDEX_PST_FILE 48
+#define INDEX_MOBILITY 96
+#define INDEX_PASSER 102
+
+#define OFFSET_PST_RANK S(-15, -11)
+#define OFFSET_PST_FILE S(-4, -5)
+#define OFFSET_MOBILITY S(-5, -1)
+#define OFFSET_PASSER S(-2, 0)
+
+int get_data(int index) {
+    auto data = DATA_STR;
+
+    return data[index] + data[index + INDEX_EG] * 0x10000 - S(32, 32);
+}
